Made a puzzle generator that gives a set of letters, and expects words.
Word list is 10k most common words on the internet, minus some dumb 3
letter words like url and fri and other random abbreviations. If this
were to become a kids game I think I'd need common words found in
children's books.

Either way, my first thought was that my son would need to find the
unique 5-letter, 4-letter, and 3-letter word based on 12 given letters.
This was HARD. The current state (configurable at the top of the file)
is to output a SINGLE 4 letter word from the most common 100 4 letter
words. My son was even struggling with this. I think I'd need to put the
words in sentences, or give pictures of 10 words with the jumbled
versions on the page. I need to think a bit more. This code probably
won't be useful for making a workstation though without a significant
change in some form.